# portable_simd version of Sub (3bpp/4bpp)